Use random string instead of autoincrementing int
The model creating event generates a 20-character random string to use
for the client id instead of an autoincrementing integer. The purpose of
this is so that users do not have insight into how many clients have
already been created. Performance is of little concern unless there are
tens of thousands (perhaps millions?) of oauth clients...unlikely.